backend/requirements.txt
fastapi
uvicorn
docker
pydantic
passlib[bcrypt]
python-jose
pyjwt
psutil
python-multipart
argon2_cffi
websockets